test-requirements.txt
vcrpy
requests_mock
freezegun
